+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -1048,15 +1048,11 @@ namespace Libplanet.Tests.Blockchain
         public void GetStatesOnUninitializedBlockChain()
         {
             bool invoked = false;
-            IReadOnlyList<IValue> values = null;
-            IValue value = null;
             var policy = new NullPolicyForGetStatesOnUninitializedBlockChain<DumbAction>(
                 c =>
                 {
                     // ReSharper disable AccessToModifiedClosure
                     // The following method calls should not throw any exceptions:
-                    values = c?.GetStates(new[] { default(Address) });
-                    value = c?.GetState(default);
                     invoked = true;
                     // ReSharper restore AccessToModifiedClosure
                 });
@@ -1086,11 +1082,7 @@ namespace Libplanet.Tests.Blockchain
                 stateStore,
                 genesisWithTx
             );
-            Assert.True(invoked);
-            Assert.NotNull(values);
-            Assert.Single(values);
-            Assert.Null(values[0]);
-            Assert.Null(value);
+            Assert.False(invoked);
         }
 
         // This is a regression test for:
@@ -2039,42 +2031,36 @@ namespace Libplanet.Tests.Blockchain
                 });
             var store = new MemoryStore();
             var stateStore = new TrieStateStore(new MemoryKeyValueStore());
-            Block<DumbAction> genesisWithTx = MineGenesis(
+
+            var genesisTx = Transaction<DumbAction>.Create(
+                0,
+                new PrivateKey(),
+                null,
+                Array.Empty<DumbAction>());
+            var genesisWithTx = MineGenesis(
                 GenesisMiner.PublicKey,
-                new[]
-                {
-                    Transaction<DumbAction>.Create(
-                        0,
-                        new PrivateKey(),
-                        null,
-                        Array.Empty<DumbAction>()
-                    ),
-                }
-            ).Evaluate(
+                new[] { genesisTx }).Evaluate(
                 privateKey: GenesisMiner,
                 blockAction: policy.BlockAction,
                 nativeTokenPredicate: policy.NativeTokens.Contains,
-                stateStore: stateStore
-            );
+                stateStore: stateStore);
 
-            Assert.Throws<TxPolicyViolationException>(() =>
-            {
-                try
-                {
-                    var chain = new BlockChain<DumbAction>(
-                        policy,
-                        new VolatileStagePolicy<DumbAction>(),
-                        store,
-                        stateStore,
-                        genesisWithTx
-                    );
-                }
-                catch (TxPolicyViolationException e)
-                {
-                    Assert.NotNull(e.InnerException);
-                    throw;
-                }
-            });
+            var chain = new BlockChain<DumbAction>(
+                policy,
+                new VolatileStagePolicy<DumbAction>(),
+                store,
+                stateStore,
+                genesisWithTx);
+
+            var blockTx = Transaction<DumbAction>.Create(
+                0,
+                new PrivateKey(),
+                null,
+                Array.Empty<DumbAction>());
+            var block = TestUtils.MineNextBlock(chain.Genesis, GenesisMiner, new[] { blockTx });
+
+            var e = Assert.Throws<TxPolicyViolationException>(() => chain.Append(block));
+            Assert.NotNull(e.InnerException);
         }
 
         private class