+++ Libplanet/Blockchain/Policies/IBlockPolicy.cs
@@ -17,6 +17,9 @@ namespace Libplanet.Blockchain.Policies
     /// such as whether a <see cref="Block{T}"/> has the right difficulty,
     /// a <see cref="Transaction{T}"/> has the right signer, etc.
     /// </para>
+    /// <para>
+    /// Note that all index dependent sub-policies are ignored for genesis <see cref="Block{T}"/>s.
+    /// </para>
     /// </summary>
     /// <typeparam name="T">An <see cref="IAction"/> type.  It should match
     /// <see cref="Block{T}"/>'s type parameter.</typeparam>
@@ -92,9 +95,15 @@ namespace Libplanet.Blockchain.Policies
             BlockChain<T> blockChain, Block<T> nextBlock);
 
         /// <summary>
+        /// <para>
         /// Determines a right <see cref="Block{T}.Difficulty"/>
         /// for a new <see cref="Block{T}"/> to be mined
         /// right after the given <paramref name="blockChain"/>.
+        /// </para>
+        /// <para>
+        /// Any specific implementation of this method should assume a genesis
+        /// <see cref="Block{T}"/> will always have zero difficulty.
+        /// </para>
         /// </summary>
         /// <param name="blockChain">Consecutive <see cref="Block{T}"/>s to be
         /// followed by a new <see cref="Block{T}"/> to be mined.</param>